Separate declaration and implementation details of public classes
